+++ packages/bsky/src/api/app/bsky/notification/listNotifications.ts
@@ -2,6 +2,7 @@ import { InvalidRequestError } from '@atproto/xrpc-server'
 import { mapDefined } from '@atproto/common'
 import { Server } from '../../../../lexicon'
 import { QueryParams } from '../../../../lexicon/types/app/bsky/notification/listNotifications'
+import { isRecord as isPostRecord } from '../../../../lexicon/types/app/bsky/feed/post'
 import AppContext from '../../../../context'
 import {
   createPipeline,
@@ -13,7 +14,7 @@ import {
 import { HydrateCtx, Hydrator } from '../../../../hydration/hydrator'
 import { Views } from '../../../../views'
 import { Notification } from '../../../../proto/bsky_pb'
-import { didFromUri } from '../../../../hydration/util'
+import { uriToDid as didFromUri } from '../../../../util/uris'
 import { clearlyBadCursor, resHeaders } from '../../../util'
 
 export default function (server: Server, ctx: AppContext) {
@@ -90,13 +91,38 @@ const hydration = async (
 const noBlockOrMutes = (
   input: RulesFnInput<Context, Params, SkeletonState>,
 ) => {
-  const { skeleton, hydration, ctx } = input
+  const { skeleton, hydration, ctx, params } = input
   skeleton.notifs = skeleton.notifs.filter((item) => {
     const did = didFromUri(item.uri)
-    return (
-      !ctx.views.viewerBlockExists(did, hydration) &&
-      !ctx.views.viewerMuteExists(did, hydration)
-    )
+    if (
+      ctx.views.viewerBlockExists(did, hydration) ||
+      ctx.views.viewerMuteExists(did, hydration)
+    ) {
+      return false
+    }
+    // Filter out hidden replies only if the viewer owns
+    // the threadgate and they hid the reply.
+    if (item.reason === 'reply') {
+      const post = hydration.posts?.get(item.uri)
+      if (post) {
+        const rootPostUri = isPostRecord(post.record)
+          ? post.record.reply?.root.uri
+          : undefined
+        const isRootPostByViewer =
+          rootPostUri && didFromUri(rootPostUri) === params.hydrateCtx?.viewer
+        const isHiddenReply = isRootPostByViewer
+          ? ctx.views.replyIsHiddenByThreadgate(
+              item.uri,
+              rootPostUri,
+              hydration,
+            )
+          : false
+        if (isHiddenReply) {
+          return false
+        }
+      }
+    }
+    return true
   })
   return skeleton
 }